Add navigation and restore original home screen
