+++ package.json
@@ -14,7 +14,11 @@
     "expo-status-bar": "~2.0.1",
     "expo-updates": "~0.27.4",
     "react": "18.3.1",
-    "react-native": "0.76.9"
+    "react-native": "0.76.9",
+    "@react-navigation/native": "^6.2.0",
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "react-native-safe-area-context": "^4.8.0",
+    "react-native-screens": "^3.31.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
