+++ src/navigation/AppNavigator.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import HomeScreen from '../screens/HomeScreen';
+import HabitsLogScreen from '../screens/HabitsLogScreen';
+
+const Tab = createBottomTabNavigator();
+
+export default function AppNavigator() {
+  return (
+    <NavigationContainer>
+      <Tab.Navigator>
+        <Tab.Screen name="Home" component={HomeScreen} />
+        <Tab.Screen name="Habits" component={HabitsLogScreen} />
+      </Tab.Navigator>
+    </NavigationContainer>
+  );
+}
