+++ src/screens/HabitsLogScreen.js
@@ -1,0 +1,69 @@
+import React, { useEffect, useState } from "react";
+import { View, Text, FlatList, StyleSheet } from "react-native";
+import { getHabits } from "../services/habits.service";
+
+export default function HabitsLogScreen() {
+  const [habits, setHabits] = useState([]);
+
+  useEffect(() => {
+    async function fetchHabits() {
+      try {
+        const data = await getHabits();
+        setHabits(data);
+      } catch (error) {
+        console.error("Failed to fetch habits", error);
+      }
+    }
+
+    fetchHabits();
+  }, []);
+
+  const renderItem = ({ item }) => (
+    <View style={styles.item}>
+      <Text style={styles.name}>{item.name}</Text>
+      <Text style={styles.date}>{item.date}</Text>
+    </View>
+  );
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Habits Log</Text>
+      <FlatList
+        data={habits}
+        keyExtractor={(item) => item.id.toString()}
+        renderItem={renderItem}
+        contentContainerStyle={styles.list}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    padding: 16,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: "bold",
+    marginBottom: 16,
+    textAlign: "center",
+  },
+  list: {
+    paddingBottom: 16,
+  },
+  item: {
+    paddingVertical: 12,
+    borderBottomWidth: 1,
+    borderColor: "#ccc",
+  },
+  name: {
+    fontSize: 16,
+    fontWeight: "bold",
+  },
+  date: {
+    fontSize: 14,
+    color: "#666",
+  },
+});
